Query: Remove obsoleted code
Part of #22549

Only touched code relating to query
