+++ src/EFCore.Cosmos/Query/Internal/CosmosQueryableMethodTranslatingExpressionVisitor.cs
@@ -233,31 +233,6 @@ namespace Microsoft.EntityFrameworkCore.Cosmos.Query.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        [Obsolete("Use overload which takes IEntityType.")]
-        protected override ShapedQueryExpression CreateShapedQueryExpression(Type elementType)
-        {
-            Check.NotNull(elementType, nameof(elementType));
-
-            var entityType = _queryCompilationContext.Model.FindEntityType(elementType);
-            var selectExpression = _sqlExpressionFactory.Select(entityType);
-
-            return new ShapedQueryExpression(
-                selectExpression,
-                new EntityShaperExpression(
-                    entityType,
-                    new ProjectionBindingExpression(
-                        selectExpression,
-                        new ProjectionMember(),
-                        typeof(ValueBuffer)),
-                    false));
-        }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
         protected override ShapedQueryExpression CreateShapedQueryExpression(IEntityType entityType)
         {
             Check.NotNull(entityType, nameof(entityType));
