+++ src/EFCore.Relational/Query/ISqlExpressionFactory.cs
@@ -24,22 +24,6 @@ namespace Microsoft.EntityFrameworkCore.Query
     public interface ISqlExpressionFactory
     {
         /// <summary>
-        ///     Gets the relational database type for a given object, throwing if no mapping is found.
-        /// </summary>
-        /// <param name="value">The object to get the mapping for.</param>
-        /// <returns>The type mapping to be used.</returns>
-        [Obsolete("Use IRelationalTypeMappingSource directly.")]
-        RelationalTypeMapping GetTypeMappingForValue(object? value);
-
-        /// <summary>
-        ///     Finds the type mapping for a given <see cref="Type" />.
-        /// </summary>
-        /// <param name="type">The CLR type.</param>
-        /// <returns>The type mapping, or <see langword="null" /> if none was found.</returns>
-        [Obsolete("Use IRelationalTypeMappingSource directly.")]
-        RelationalTypeMapping? FindMapping(Type type);
-
-        /// <summary>
         ///     Applies type mapping to the given <see cref="SqlExpression" />.
         /// </summary>
         /// <param name="sqlExpression">A SQL expression to apply type mapping.</param>
@@ -294,15 +278,6 @@ namespace Microsoft.EntityFrameworkCore.Query
         /// </summary>
         /// <param name="operand">An expression to compare with <see cref="CaseWhenClause.Test" /> in <paramref name="whenClauses" />.</param>
         /// <param name="whenClauses">A list of <see cref="CaseWhenClause" /> to compare and get result from.</param>
-        /// <returns>An expression representing a CASE statement in a SQL tree.</returns>
-        [Obsolete("Use overload which takes IReadOnlyList instead of params")]
-        CaseExpression Case(SqlExpression operand, params CaseWhenClause[] whenClauses);
-
-        /// <summary>
-        ///     Creates a new <see cref="CaseExpression" /> which represent a CASE statement in a SQL tree.
-        /// </summary>
-        /// <param name="operand">An expression to compare with <see cref="CaseWhenClause.Test" /> in <paramref name="whenClauses" />.</param>
-        /// <param name="whenClauses">A list of <see cref="CaseWhenClause" /> to compare and get result from.</param>
         /// <param name="elseResult">A value to return if no <paramref name="whenClauses" /> matches, if any.</param>
         /// <returns>An expression representing a CASE statement in a SQL tree.</returns>
         CaseExpression Case(
@@ -317,99 +292,6 @@ namespace Microsoft.EntityFrameworkCore.Query
         /// <param name="elseResult">A value to return if no <paramref name="whenClauses" /> matches, if any.</param>
         /// <returns>An expression representing a CASE statement in a SQL tree.</returns>
         CaseExpression Case(IReadOnlyList<CaseWhenClause> whenClauses, SqlExpression? elseResult);
-
-        /// <summary>
-        ///     Creates a new <see cref="SqlFunctionExpression" /> which represents a function call in a SQL tree.
-        /// </summary>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="arguments">The arguments of the function.</param>
-        /// <param name="returnType">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        /// <returns>An expression representing a function call in a SQL tree.</returns>
-        [Obsolete("Use overload that explicitly specifies value for 'argumentsPropagateNullability' argument.")]
-        SqlFunctionExpression Function(
-            string name,
-            IEnumerable<SqlExpression> arguments,
-            Type returnType,
-            RelationalTypeMapping? typeMapping = null);
-
-        /// <summary>
-        ///     Creates a new <see cref="SqlFunctionExpression" /> which represents a function call in a SQL tree.
-        /// </summary>
-        /// <param name="schema">The schema in which the function is defined.</param>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="arguments">The arguments of the function.</param>
-        /// <param name="returnType">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        /// <returns>An expression representing a function call in a SQL tree.</returns>
-        [Obsolete("Use overload that explicitly specifies value for 'argumentsPropagateNullability' argument.")]
-        SqlFunctionExpression Function(
-            string? schema,
-            string name,
-            IEnumerable<SqlExpression> arguments,
-            Type returnType,
-            RelationalTypeMapping? typeMapping = null);
-
-        /// <summary>
-        ///     Creates a new <see cref="SqlFunctionExpression" /> which represents a function call in a SQL tree.
-        /// </summary>
-        /// <param name="instance">An expression on which the function is applied.</param>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="arguments">The arguments of the function.</param>
-        /// <param name="returnType">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        /// <returns>An expression representing a function call in a SQL tree.</returns>
-        [Obsolete(
-            "Use overload that explicitly specifies value for 'instancePropagatesNullability' and 'argumentsPropagateNullability' arguments.")]
-        SqlFunctionExpression Function(
-            SqlExpression instance,
-            string name,
-            IEnumerable<SqlExpression> arguments,
-            Type returnType,
-            RelationalTypeMapping? typeMapping = null);
-
-        /// <summary>
-        ///     Creates a new <see cref="SqlFunctionExpression" /> which represents a function call in a SQL tree.
-        /// </summary>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="returnType">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        /// <returns>An expression representing a function call in a SQL tree.</returns>
-        [Obsolete("Use NiladicFunction method.")]
-        SqlFunctionExpression Function(
-            string name,
-            Type returnType,
-            RelationalTypeMapping? typeMapping = null);
-
-        /// <summary>
-        ///     Creates a new <see cref="SqlFunctionExpression" /> which represents a function call in a SQL tree.
-        /// </summary>
-        /// <param name="schema">The schema in which the function is defined.</param>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="returnType">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        /// <returns>An expression representing a function call in a SQL tree.</returns>
-        [Obsolete("Use NiladicFunction method.")]
-        SqlFunctionExpression Function(
-            string schema,
-            string name,
-            Type returnType,
-            RelationalTypeMapping? typeMapping = null);
-
-        /// <summary>
-        ///     Creates a new <see cref="SqlFunctionExpression" /> which represents a function call in a SQL tree.
-        /// </summary>
-        /// <param name="instance">An expression on which the function is applied.</param>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="returnType">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        /// <returns>An expression representing a function call in a SQL tree.</returns>
-        [Obsolete("Use NiladicFunction method.")]
-        SqlFunctionExpression Function(
-            SqlExpression instance,
-            string name,
-            Type returnType,
-            RelationalTypeMapping? typeMapping = null);
 
         /// <summary>
         ///     Creates a new <see cref="SqlFunctionExpression" /> which represents a function call in a SQL tree.
@@ -593,16 +475,5 @@ namespace Microsoft.EntityFrameworkCore.Query
         /// <param name="tableExpressionBase">A table source to project from.</param>
         /// <returns>An expression representing a SELECT in a SQL tree.</returns>
         SelectExpression Select(IEntityType entityType, TableExpressionBase tableExpressionBase);
-
-        /// <summary>
-        ///     Creates a new <see cref="SelectExpression" /> which represents a SELECT in a SQL tree projecting an entity type from
-        ///     a table source created using a custom SQL.
-        /// </summary>
-        /// <param name="entityType">An entity type to project.</param>
-        /// <param name="sql">A custom SQL for the table source.</param>
-        /// <param name="sqlArguments">An expression representing parameters passed to the custom SQL.</param>
-        /// <returns>An expression representing a SELECT in a SQL tree.</returns>
-        [Obsolete("Use overload which takes TableExpressionBase by passing FromSqlExpression directly.")]
-        SelectExpression Select(IEntityType entityType, string sql, Expression sqlArguments);
     }
 }
