+++ src/EFCore.Relational/Query/RelationalCollectionShaperExpression.cs
@@ -25,75 +25,6 @@ namespace Microsoft.EntityFrameworkCore.Query
         /// <summary>
         ///     Creates a new instance of the <see cref="RelationalCollectionShaperExpression" /> class.
         /// </summary>
-        /// <param name="collectionId">A unique id for the collection being shaped.</param>
-        /// <param name="parentIdentifier">An identifier for the parent element.</param>
-        /// <param name="outerIdentifier">An identifier for the outer element.</param>
-        /// <param name="selfIdentifier">An identifier for the element in the collection.</param>
-        /// <param name="innerShaper">An expression used to create individual elements of the collection.</param>
-        /// <param name="navigation">A navigation associated with this collection, if any.</param>
-        /// <param name="elementType">The clr type of individual elements in the collection.</param>
-        [Obsolete("Use ctor which takes value comparers.")]
-        public RelationalCollectionShaperExpression(
-            int collectionId,
-            Expression parentIdentifier,
-            Expression outerIdentifier,
-            Expression selfIdentifier,
-            Expression innerShaper,
-            INavigation? navigation,
-            Type elementType)
-            : this(
-                collectionId, parentIdentifier, outerIdentifier, selfIdentifier,
-                null, null, null, innerShaper, navigation, elementType)
-        {
-        }
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="RelationalCollectionShaperExpression" /> class.
-        /// </summary>
-        /// <param name="collectionId">A unique id for the collection being shaped.</param>
-        /// <param name="parentIdentifier">An identifier for the parent element.</param>
-        /// <param name="outerIdentifier">An identifier for the outer element.</param>
-        /// <param name="selfIdentifier">An identifier for the element in the collection.</param>
-        /// <param name="parentIdentifierValueComparers">A list of value comparers to compare parent identifier.</param>
-        /// <param name="outerIdentifierValueComparers">A list of value comparers to compare outer identifier.</param>
-        /// <param name="selfIdentifierValueComparers">A list of value comparers to compare self identifier.</param>
-        /// <param name="innerShaper">An expression used to create individual elements of the collection.</param>
-        /// <param name="navigation">A navigation associated with this collection, if any.</param>
-        /// <param name="elementType">The clr type of individual elements in the collection.</param>
-        [Obsolete("Use ctor without collectionId")]
-        public RelationalCollectionShaperExpression(
-            int collectionId,
-            Expression parentIdentifier,
-            Expression outerIdentifier,
-            Expression selfIdentifier,
-            IReadOnlyList<ValueComparer>? parentIdentifierValueComparers,
-            IReadOnlyList<ValueComparer>? outerIdentifierValueComparers,
-            IReadOnlyList<ValueComparer>? selfIdentifierValueComparers,
-            Expression innerShaper,
-            INavigationBase? navigation,
-            Type elementType)
-        {
-            Check.NotNull(parentIdentifier, nameof(parentIdentifier));
-            Check.NotNull(outerIdentifier, nameof(outerIdentifier));
-            Check.NotNull(selfIdentifier, nameof(selfIdentifier));
-            Check.NotNull(innerShaper, nameof(innerShaper));
-            Check.NotNull(elementType, nameof(elementType));
-
-            CollectionId = collectionId;
-            ParentIdentifier = parentIdentifier;
-            OuterIdentifier = outerIdentifier;
-            SelfIdentifier = selfIdentifier;
-            ParentIdentifierValueComparers = parentIdentifierValueComparers;
-            OuterIdentifierValueComparers = outerIdentifierValueComparers;
-            SelfIdentifierValueComparers = selfIdentifierValueComparers;
-            InnerShaper = innerShaper;
-            Navigation = navigation;
-            ElementType = elementType;
-        }
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="RelationalCollectionShaperExpression" /> class.
-        /// </summary>
         /// <param name="parentIdentifier">An identifier for the parent element.</param>
         /// <param name="outerIdentifier">An identifier for the outer element.</param>
         /// <param name="selfIdentifier">An identifier for the element in the collection.</param>
@@ -130,12 +61,6 @@ namespace Microsoft.EntityFrameworkCore.Query
             Navigation = navigation;
             ElementType = elementType;
         }
-
-        /// <summary>
-        ///     A unique id for this collection shaper.
-        /// </summary>
-        [Obsolete("CollectionId are not stored in shaper anymore. Shaper compiler assigns it as needed.")]
-        public virtual int CollectionId { get; }
 
         /// <summary>
         ///     The identifier for the parent element.
