+++ src/EFCore.Relational/Query/RelationalSplitCollectionShaperExpression.cs
@@ -26,44 +26,6 @@ namespace Microsoft.EntityFrameworkCore.Query
         /// <summary>
         ///     Creates a new instance of the <see cref="RelationalCollectionShaperExpression" /> class.
         /// </summary>
-        /// <param name="collectionId">A unique id for the collection being shaped.</param>
-        /// <param name="parentIdentifier">An identifier for the parent element.</param>
-        /// <param name="childIdentifier">An identifier for the child element.</param>
-        /// <param name="identifierValueComparers">A list of value comparers to compare identifiers.</param>
-        /// <param name="selectExpression">A SQL query to get values for this collection from database.</param>
-        /// <param name="innerShaper">An expression used to create individual elements of the collection.</param>
-        /// <param name="navigation">A navigation associated with this collection, if any.</param>
-        /// <param name="elementType">The clr type of individual elements in the collection.</param>
-        [Obsolete("Use ctor without collectionId")]
-        public RelationalSplitCollectionShaperExpression(
-            int collectionId,
-            Expression parentIdentifier,
-            Expression childIdentifier,
-            IReadOnlyList<ValueComparer> identifierValueComparers,
-            SelectExpression selectExpression,
-            Expression innerShaper,
-            INavigationBase? navigation,
-            Type elementType)
-        {
-            Check.NotNull(parentIdentifier, nameof(parentIdentifier));
-            Check.NotNull(childIdentifier, nameof(childIdentifier));
-            Check.NotEmpty(identifierValueComparers, nameof(identifierValueComparers));
-            Check.NotNull(innerShaper, nameof(innerShaper));
-            Check.NotNull(elementType, nameof(elementType));
-
-            CollectionId = collectionId;
-            ParentIdentifier = parentIdentifier;
-            ChildIdentifier = childIdentifier;
-            IdentifierValueComparers = identifierValueComparers;
-            SelectExpression = selectExpression;
-            InnerShaper = innerShaper;
-            Navigation = navigation;
-            ElementType = elementType;
-        }
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="RelationalCollectionShaperExpression" /> class.
-        /// </summary>
         /// <param name="parentIdentifier">An identifier for the parent element.</param>
         /// <param name="childIdentifier">An identifier for the child element.</param>
         /// <param name="identifierValueComparers">A list of value comparers to compare identifiers.</param>
@@ -94,12 +56,6 @@ namespace Microsoft.EntityFrameworkCore.Query
             Navigation = navigation;
             ElementType = elementType;
         }
-
-        /// <summary>
-        ///     A unique id for this collection shaper.
-        /// </summary>
-        [Obsolete("CollectionId are not stored in shaper anymore. Shaper compiler assigns it as needed.")]
-        public virtual int CollectionId { get; }
 
         /// <summary>
         ///     The identifier for the parent element.
