+++ src/EFCore.Relational/Query/SqlExpressions/SqlFunctionExpression.cs
@@ -410,115 +410,5 @@ namespace Microsoft.EntityFrameworkCore.Query.SqlExpressions
 
             return hash.ToHashCode();
         }
-
-        #region ObsoleteMethods
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="SqlFunctionExpression" /> class.
-        /// </summary>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="type">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        [Obsolete("Use new SqlFunctionExpression(...) with appropriate arguments.")]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static SqlFunctionExpression CreateNiladic(
-            string name,
-            Type type,
-            RelationalTypeMapping? typeMapping)
-            => new(name, nullable: true, type, typeMapping);
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="SqlFunctionExpression" /> class.
-        /// </summary>
-        /// <param name="schema">The schema in which the function is defined.</param>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="type">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        [Obsolete("Use new SqlFunctionExpression(...) with appropriate arguments.")]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static SqlFunctionExpression CreateNiladic(
-            string schema,
-            string name,
-            Type type,
-            RelationalTypeMapping? typeMapping)
-            => new(schema, name, nullable: true, type, typeMapping);
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="SqlFunctionExpression" /> class.
-        /// </summary>
-        /// <param name="instance">An expression on which the function is defined.</param>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="type">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        [Obsolete("Use new SqlFunctionExpression(...) with appropriate arguments.")]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static SqlFunctionExpression CreateNiladic(
-            SqlExpression instance,
-            string name,
-            Type type,
-            RelationalTypeMapping? typeMapping)
-            => new(instance, name, nullable: true, instancePropagatesNullability: false, type, typeMapping);
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="SqlFunctionExpression" /> class.
-        /// </summary>
-        /// <param name="instance">An expression on which the function is applied.</param>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="arguments">The arguments of the function.</param>
-        /// <param name="type">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        [Obsolete("Use new SqlFunctionExpression(...) with appropriate arguments.")]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static SqlFunctionExpression Create(
-            SqlExpression instance,
-            string name,
-            IEnumerable<SqlExpression> arguments,
-            Type type,
-            RelationalTypeMapping? typeMapping)
-            => new(
-                instance,
-                name,
-                arguments,
-                nullable: true,
-                instancePropagatesNullability: false,
-                argumentsPropagateNullability: arguments.Select(a => false),
-                type,
-                typeMapping);
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="SqlFunctionExpression" /> class.
-        /// </summary>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="arguments">The arguments of the function.</param>
-        /// <param name="type">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        [Obsolete("Use new SqlFunctionExpression(...) with appropriate arguments.")]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static SqlFunctionExpression Create(
-            string name,
-            IEnumerable<SqlExpression> arguments,
-            Type type,
-            RelationalTypeMapping? typeMapping)
-            => new(name, arguments, nullable: true, argumentsPropagateNullability: arguments.Select(a => false), type, typeMapping);
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="SqlFunctionExpression" /> class.
-        /// </summary>
-        /// <param name="schema">The schema in which the function is defined.</param>
-        /// <param name="name">The name of the function.</param>
-        /// <param name="arguments">The arguments of the function.</param>
-        /// <param name="type">The <see cref="Type" /> of the expression.</param>
-        /// <param name="typeMapping">The <see cref="RelationalTypeMapping" /> associated with the expression.</param>
-        [Obsolete("Use new SqlFunctionExpression(...) with appropriate arguments.")]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static SqlFunctionExpression Create(
-            string? schema,
-            string name,
-            IEnumerable<SqlExpression> arguments,
-            Type type,
-            RelationalTypeMapping? typeMapping)
-            => new(schema, name, arguments, nullable: true, argumentsPropagateNullability: arguments.Select(a => false), type, typeMapping);
-
-        #endregion
     }
 }
