+++ src/EFCore.Relational/Storage/ReaderColumn.cs
@@ -32,18 +32,6 @@ namespace Microsoft.EntityFrameworkCore.Storage
         /// <param name="type">The CLR type of the column.</param>
         /// <param name="nullable">A value indicating if the column is nullable.</param>
         /// <param name="name">The name of the column.</param>
-        [Obsolete("Use constructor which also takes IPropertyBase.")]
-        protected ReaderColumn(Type type, bool nullable, string? name)
-            : this(type, nullable, name, null)
-        {
-        }
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="ReaderColumn" /> class.
-        /// </summary>
-        /// <param name="type">The CLR type of the column.</param>
-        /// <param name="nullable">A value indicating if the column is nullable.</param>
-        /// <param name="name">The name of the column.</param>
         /// <param name="property">The property being read if any, null otherwise.</param>
         protected ReaderColumn(Type type, bool nullable, string? name, IPropertyBase? property)
         {
@@ -72,20 +60,6 @@ namespace Microsoft.EntityFrameworkCore.Storage
         ///     The property being read if any, null otherwise.
         /// </summary>
         public virtual IPropertyBase? Property { get; }
-
-        /// <summary>
-        ///     Creates an instance of <see cref="ReaderColumn{T}" />.
-        /// </summary>
-        /// <param name="type">The type of the column.</param>
-        /// <param name="nullable">Whether the column can contain <see langword="null" /> values.</param>
-        /// <param name="columnName">The column name if it is used to access the column values, <see langword="null" /> otherwise.</param>
-        /// <param name="readFunc">
-        ///     A <see cref="T:System.Func{DbDataReader, Int32[], T}" /> used to get the field value for this column.
-        /// </param>
-        /// <returns>An instance of <see cref="ReaderColumn{T}" />.</returns>
-        [Obsolete("Use method which also takes IPropertyBase.")]
-        public static ReaderColumn Create(Type type, bool nullable, string? columnName, object readFunc)
-            => (ReaderColumn)GetConstructor(type).Invoke(new[] { nullable, columnName, readFunc });
 
         /// <summary>
         ///     Creates an instance of <see cref="ReaderColumn{T}" />.
