+++ src/EFCore.Relational/Storage/ReaderColumn`.cs
@@ -27,18 +27,6 @@ namespace Microsoft.EntityFrameworkCore.Storage
         /// </summary>
         /// <param name="nullable">A value indicating if the column is nullable.</param>
         /// <param name="name">The name of the column.</param>
-        /// <param name="getFieldValue">A function to get field value for the column from the reader.</param>
-        [Obsolete("Use constructor which also takes IPropertyBase.")]
-        public ReaderColumn(bool nullable, string? name, Func<DbDataReader, int[], T> getFieldValue)
-            : this(nullable, name, property: null, getFieldValue)
-        {
-        }
-
-        /// <summary>
-        ///     Creates a new instance of the <see cref="ReaderColumn{T}" /> class.
-        /// </summary>
-        /// <param name="nullable">A value indicating if the column is nullable.</param>
-        /// <param name="name">The name of the column.</param>
         /// <param name="property">The property being read if any, null otherwise.</param>
         /// <param name="getFieldValue">A function to get field value for the column from the reader.</param>
         public ReaderColumn(
